requirements.txt
cryptography==2.9.2
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pkg-resources==0.0.0
python-jose[cryptography]==3.1.0
Werkzeug==0.14.1
